separated training-related functions into it's own file
